% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{contract_fields}
\alias{contract_fields}
\title{Field names of a contract.}
\usage{
contract_fields(contract)
}
\arguments{
\item{contract}{A contract (list of specs or tibble).}
}
\value{
Character vector of field names, in contract order.
}
\description{
Field names of a contract.
}
