% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{delete_mapping_profile}
\alias{delete_mapping_profile}
\title{Delete a mapping profile}
\usage{
delete_mapping_profile(
  name_or_path,
  dir = getOption("tritonIngest.profiles_dir")
)
}
\arguments{
\item{name_or_path}{Profile name/path.}

\item{dir}{Profiles directory.}
}
\value{
\code{TRUE} if removed.
}
\description{
Delete a mapping profile
}
