% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Resolve the cache directory}
\usage{
cache_dir(dir = getOption("tritonIngest.cache_dir"), create = TRUE)
}
\arguments{
\item{dir}{Directory to store cache files in.}

\item{create}{Create it when missing.}
}
\value{
Normalized absolute path.
}
\description{
Resolve the cache directory
}
