% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{list_mapping_profiles}
\alias{list_mapping_profiles}
\title{List mapping profiles}
\usage{
list_mapping_profiles(dir = getOption("tritonIngest.profiles_dir"))
}
\arguments{
\item{dir}{Profiles directory.}
}
\value{
Tibble with name, schema, file, and modified time.
}
\description{
List mapping profiles
}
